requirement.txt
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.2
charset-normalizer==2.0.6
colorama==0.4.3
cryptography==3.0
decorator==4.4.2
distlib==0.3.0
et-xmlfile==1.1.0
filelock==3.0.12
idna==3.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.21.2
openai==0.10.5
openpyxl==3.0.9
pandas==1.3.3
pandas-stubs==1.2.0.19
pycparser==2.20
pylint==2.5.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.3
requests==2.26.0
six==1.15.0
toml==0.10.1
tornado==6.1
tqdm==4.62.3
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.7
virtualenv==20.0.23
wrapt==1.12.1
